+++ src/routes/SQLConsole/SQLConsole.tsx
@@ -1,18 +1,19 @@
-import { useEffect, useState } from 'react';
-import { Heading, Loader, SQLEditor, SQLHistory, SQLResults } from 'components';
-import { useGCContext } from 'contexts';
+import { useEffect, useRef, useState } from 'react';
+import { Panel, PanelGroup, PanelResizeHandle } from 'react-resizable-panels';
+import { Loader, NoDataView, SQLEditor, SQLHistory, SQLResults } from 'components';
 import useExecuteSql from 'hooks/useExecuteSql';
 import { QueryResults } from 'types/query';
 
 function SQLConsole() {
   const executeSql = useExecuteSql();
-  const { headings } = useGCContext();
   const specifiedQuery = new URLSearchParams(location.search).get('q');
   const [results, setResults] = useState<QueryResults | undefined>(undefined);
   const [running, setRunning] = useState(false);
   const [showHistory, setShowHistory] = useState(false);
   const [currentQuery, setCurrentQuery] = useState<string | null>(null);
   const [rerender, setRerender] = useState(false);
+  const outerHeight = useRef(null);
+  const innerHeight = useRef(null);
 
   // if a query is specified in the URL, place it in the editor
   useEffect(() => {
@@ -56,23 +57,30 @@ function SQLConsole() {
   };
 
   return (
-    <div>
-      {headings && (
-        <Heading level="h1" className="mb-2">
-          Console
-        </Heading>
-      )}
-      <SQLEditor
-        onExecute={execute}
-        results={results}
-        localStorageKey={LOCAL_STORAGE_KEY}
-        value={currentQuery}
-        setShowHistory={setShowHistory}
-      />
-      <div className="mt-4">
-        {running && <Loader />}
-        {!running && <SQLResults results={results} />}
-      </div>
+    <>
+      <PanelGroup direction="vertical" ref={outerHeight}>
+        <Panel defaultSize={40} minSize={5} ref={innerHeight}>
+          <div className="h-full overflow-auto">
+            <SQLEditor
+              onExecute={execute}
+              localStorageKey={LOCAL_STORAGE_KEY}
+              results={results}
+              setShowHistory={setShowHistory}
+              value={currentQuery}
+            />
+          </div>
+        </Panel>
+        <PanelResizeHandle className="flex h-1 justify-center bg-neutral-200 hover:bg-crate-blue">
+          <div className="h-full w-10 bg-crate-blue" />
+        </PanelResizeHandle>
+        <Panel minSize={5}>
+          <div className="h-full overflow-auto">
+            {running && <Loader />}
+            {!running && <SQLResults results={results} />}
+            {!running && !results && <NoDataView />}
+          </div>
+        </Panel>
+      </PanelGroup>
       <SQLHistory
         history={history}
         showHistory={showHistory}
@@ -81,7 +89,7 @@ function SQLConsole() {
         displayHistoryItem={displayHistoryItem}
         setShowHistory={setShowHistory}
       />
-    </div>
+    </>
   );
 }